renamed forms to WFOForms
