+++ apps/wfo-ui/pages/forms.tsx
@@ -1,8 +1,9 @@
 import { EuiHorizontalRule, EuiPageHeader, EuiSpacer } from '@elastic/eui';
 import React from 'react';
-import CreateForm from '@orchestrator-ui/orchestrator-ui-components/src/components/forms/CreateForm';
+import CreateForm from '@orchestrator-ui/orchestrator-ui-components/src/components/WFOForms/CreateForm';
 
 export function FormsPage() {
+    /* eslint-disable  @typescript-eslint/no-explicit-any */
     const handleSubmit = (userInputs: any) => {
         console.log('Submitted: ', userInputs);
     };
